+++ RELEASE_NOTES.md
@@ -1,4 +1,12 @@
-#### 5.176.1 - 2018-08-02
+#### 5.177.1 - 2018-08-30
+* TEMPORARY WORKAROUND: Do not reference Microsoft.Azure.WebJobs.Script.ExtensionsMetadataGenerator - https://github.com/fsprojects/Paket/issues/3345
+
+#### 5.177.0 - 2018-08-30
+* BUGFIX: fix 'Value cannot be null' with TFS feed - https://github.com/fsprojects/Paket/pull/3341
+* BUGFIX: Fixed authentication problem in netcore (fake 5) - https://github.com/fsprojects/Paket/pull/3342
+* BUGFIX: Fixed 'Value cannot be null' with TFS feed - https://github.com/fsprojects/Paket/pull/3341
+* BUGFIX: netcoreapp2.0 packages are compatibile with netcoreapp2.1 - https://github.com/fsprojects/Paket/pull/3336
+* BUGFIX: Paket uses ToLowerInvariant instead of ToLower when forming registrationUrl - https://github.com/fsprojects/Paket/pull/3330
 * BUGFIX: Removed faulty creation of directories during `generate-load-scripts` - https://github.com/fsprojects/Paket/pull/3319
 
 #### 5.176.0 - 2018-07-31